requirements.txt
asgiref==3.8.1
attrs==23.2.0
dj-database-url==2.2.0
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
importlib_metadata==7.1.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
packaging==24.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
zipp==3.18.1
